SyntaxInterface

Documentation for SyntaxInterface.

SyntaxInterface.arguments

SyntaxInterface.arity

SyntaxInterface.is_operation

SyntaxInterface.istree

SyntaxInterface.node_count

SyntaxInterface.operation

SyntaxInterface.similarterm

SyntaxInterface.similarterm

SyntaxInterface.arguments — Function

arguments(x)

If x is a term as defined by istree(x), arguments(x) returns the arguments of x.

operation(x)(arguments(x)...) should produce a node equivalent to x.

SyntaxInterface.arity — Method

arity(x)

Returns the number of arguments of x. Implicitly defined if arguments(x) is defined.

SyntaxInterface.is_operation — Method

is_operation(f)

Returns a single argument anonymous function predicate, that returns true if and only if the argument to the predicate satisfies istree and operation(x) == f

SyntaxInterface.istree — Method

istree(x)

Returns true if x is a term. If true, operation, arguments must also be defined for x appropriately.

SyntaxInterface.node_count — Method

node_count(t)

Count the nodes in a symbolic expression tree satisfying istree and arguments.

SyntaxInterface.operation — Function

operation(x)

If x is a term as defined by istree(x), operation(x) returns the head of x.

operation(x)(arguments(x)...) should produce a node equivalent to x.

SyntaxInterface.similarterm — Method

similarterm(x::T, head, args)

Returns a term that is in the same closure of nodes with type T, with head as the head and args as the arguments. By default this will execute head(args...).

Do not overload directly, overload similarterm(::Type{<:T}, ...) instead.

SyntaxInterface.similarterm — Method

similarterm(::Type{T}, head, args)

Returns a term that is in the same closure of nodes with type T, with head as the head and args as the arguments. By default this will execute head(args...).